Check PR comments for the merge trigger
Search PR comments for the merge trigger. PR comments are retrieved
via a new GET request to the activities API that must be serialized
after the request to get all PR's, which adds some delay.

Only PR comments written by the authenticated user are searched. For the
user's own PR's the username is parsed from the PR data itself so the
operation is fairly cheap. For PR's the user has approved, the username
must be retrieved by a new GET request. However, this request can be
made concurrently for virtually no additional cost.